remove bad validation code
because Coins.Sort() sorts coins in-place,
this if-statement will always return true and as a side effect
it'll modify metaData.ReserveCoins.
since this validation is for checking the Coins are sorted already,
if-statement below it will do that job instead.

WARNING: removing this if-statement will also remove the side-effect
mentioned above(sorting Coins).