LM-2: Changing Swap Fee to Dynamic
This pull request is created as another idea for resolving issue #43.
The idea here is to have swap fee rate defined in the pool and allowing
this rate to be changed by pool governor account.

Pool governor account is the account that has the most funds in the pool.
Initially, pool governor is the account that creates the pool.
Pool governor is tracked on deposit and withdrawal actions.
After pool creation, other account can become governor by depositing
more funds to the pool than current governor has.
When current pool governor withdraws enough funds from the pool other
account with most funds in the pool becomes a new governor.